backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
openai==1.3.5
google-generativeai==0.3.2
PyPDF2==3.0.1
python-multipart==0.0.6
pydantic==2.5.0
reportlab==4.0.7